Moves ui form elements from initiative creation to ui directory
