+++ src/pages/monitoring/outlets/initiativesAdmin/InitiativeCard.tsx
@@ -19,7 +19,6 @@ import {
   INITIATIVE_LOCATIONS_MIN_AMOUNT,
 } from "@config/monitoring";
 
-import { uiText } from "pages/monitoring/outlets/initiativesAdmin/layout/uiText";
 import {
   RoleInInitiative,
   type UserItem,
@@ -31,20 +30,21 @@ import type {
   InitiativeFullInfo,
   LocationObj,
 } from "pages/monitoring/types/initiative";
-import type { CardInfoGrouped } from "pages/monitoring/outlets/initiativesAdmin/types/initiativeData";
+import type { CardInfoGrouped } from "pages/monitoring/types/initiativeData";
 import {
   getInitiative,
   isMonitoringAPIError,
   monitoringAPI,
 } from "pages/monitoring/api/monitoringAPI";
-import { InitiativeStatusDialog } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/InitiativeStatusDialog";
-import { FormListUpdater } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/FormListUpdater";
-import { LocationInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/LocationInput";
-import { makeLocationObj } from "pages/monitoring/outlets/initiativesAdmin/utils/builders";
-import { UsersInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/UsersInput";
-import { ContactInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/ContactInput";
-import { GeneralInfoUpdater } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/GeneralInfoUpdater";
-import { ImagesUpdater } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/ImagesUpdater";
+import { makeLocationObj } from "pages/monitoring/ui/initiativesAdmin/utils/builders";
+import { InitiativeStatusDialog } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/InitiativeStatusDialog";
+import { FormListUpdater } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/FormListUpdater";
+import { LocationInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/LocationInput";
+import { UsersInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/UsersInput";
+import { ContactInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/ContactInput";
+import { GeneralInfoUpdater } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/GeneralInfoUpdater";
+import { ImagesUpdater } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/ImagesUpdater";
+import { uiText } from "pages/monitoring/outlets/initiativesAdmin/layout/uiText";
 
 export type InitiativeCtxType = {
   initiative: CardInfoGrouped | null;
