+++ src/pages/monitoring/outlets/initiativesAdmin/InitiativeDataForm.tsx
@@ -9,7 +9,6 @@ import {
   INITIATIVE_LEADERS_MAX_AMOUNT,
 } from "@config/monitoring";
 
-import { uiText } from "pages/monitoring/outlets/initiativesAdmin/layout/uiText";
 import type { UserItem } from "pages/monitoring/types/catalog";
 import type {
   InitiativeContact,
@@ -19,25 +18,26 @@ import type {
 import type {
   InitiativeDataForm,
   InitiativeDataFormErr,
-} from "pages/monitoring/outlets/initiativesAdmin/types/initiativeData";
+} from "pages/monitoring/types/initiativeData";
 import {
   isMonitoringAPIError,
   monitoringAPI,
   uploadImages,
 } from "pages/monitoring/api/monitoringAPI";
-import { GeneralInfoInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/GeneralInfo";
-import { FormListManager } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/FormListManager";
-import { LocationInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/LocationInput";
-import { ContactInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/ContactInput";
-import { UsersInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/UsersInput";
-import { validateFormClient } from "pages/monitoring/outlets/initiativesAdmin/utils/validateFormClient";
-import { newInitiativeValidations } from "pages/monitoring/outlets/initiativesAdmin/utils/formClientValidations";
-import { ImagesInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/ImagesInput";
+import { validateFormClient } from "pages/monitoring/ui/initiativesAdmin/utils/validateFormClient";
+import { newInitiativeValidations } from "pages/monitoring/ui/initiativesAdmin/utils/formClientValidations";
+import { GeneralInfoInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/GeneralInfo";
+import { FormListManager } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/FormListManager";
+import { LocationInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/LocationInput";
+import { ContactInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/ContactInput";
+import { UsersInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/UsersInput";
+import { ImagesInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/ImagesInput";
 import {
   makeInitialInfo,
   setFormField,
-} from "pages/monitoring/outlets/initiativesAdmin/utils/formObjectUpdate";
-import { fetchAndMakeLocationObj } from "pages/monitoring/outlets/initiativesAdmin/utils/builders";
+} from "pages/monitoring/ui/initiativesAdmin/utils/formObjectUpdate";
+import { fetchAndMakeLocationObj } from "pages/monitoring/ui/initiativesAdmin/utils/builders";
+import { uiText } from "pages/monitoring/outlets/initiativesAdmin/layout/uiText";
 
 export function InitiativeDataForm({ onSuccess }: { onSuccess: () => void }) {
   const [formID, setformID] = useState(0);
