+++ src/pages/monitoring/ui/initiativesAdmin/initiativeCard/ImagePreview.tsx
