+++ src/pages/monitoring/ui/initiativesAdmin/initiativeCard/ImagesUpdater.tsx
@@ -12,18 +12,18 @@ import { cn } from "@ui/shadCN/lib/utils";
 import type {
   CardInfoGrouped,
   InitiativeDataFormErr,
-} from "pages/monitoring/outlets/initiativesAdmin/types/initiativeData";
-import { uiText } from "pages/monitoring/outlets/initiativesAdmin/layout/uiText";
+} from "pages/monitoring/types/initiativeData";
 import {
   InitiativeCtx,
   type InitiativeCtxType,
 } from "pages/monitoring/outlets/initiativesAdmin/InitiativeCard";
 import { ErrorsList } from "@ui/LabelingWithErrors";
-import { EditModeButton } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/EditModeButton";
-import { ImagePreview } from "pages/monitoring/outlets/initiativesAdmin/initiativeCard/ImagePreview";
+import { EditModeButton } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/EditModeButton";
+import { ImagePreview } from "pages/monitoring/ui/initiativesAdmin/initiativeCard/ImagePreview";
+import { ImagesInput } from "pages/monitoring/ui/initiativesAdmin/initiativeDataForm/ImagesInput";
 import { Button } from "@ui/shadCN/component/button";
 import { uploadImages } from "pages/monitoring/api/monitoringAPI";
-import { ImagesInput } from "pages/monitoring/outlets/initiativesAdmin/initiativeDataForm/ImagesInput";
+import { uiText } from "pages/monitoring/ui/initiativesAdmin/layout/uiText";
 
 type ImagesUpdaterProps = {
   title: string;
